+++ src/controllers/trading-competition.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get } from "@nestjs/common";
 import { ApiTags } from "@nestjs/swagger";
 import { OnchainDataService } from "@services/onchain-data.service";
-import { IRanking, IRankingItemFormatted } from "../interfaces/tradingCompetition";
+import { IRankingItemFormatted } from "../interfaces/trading-competition";
 
 @Controller({ path: "trading-competition", version: "1" })
 @ApiTags("trading-competition")
@@ -9,8 +9,8 @@ export class TradingCompetitionController
 {
     public constructor(private readonly onchainDataService: OnchainDataService) {}
 
-    @Get()
-    public getTradingCompetitionRanking(): IRankingItemFormatted[]
+    @Get("ranking")
+    public getRanking(): IRankingItemFormatted[]
     {
         return this.onchainDataService.getTradingCompetitionRanking();
     }
