+++ src/interfaces/trading-competition.ts
