+++ src/main.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-import { ValidationPipe, VersioningType, VERSION_NEUTRAL } from "@nestjs/common";
+import { Logger, ValidationPipe, VersioningType, VERSION_NEUTRAL } from "@nestjs/common";
 import { ConfigService } from "@nestjs/config";
 import { NestFactory } from "@nestjs/core";
 import { FastifyAdapter, NestFastifyApplication } from "@nestjs/platform-fastify";
@@ -15,6 +15,7 @@ async function bootstrap()
         new FastifyAdapter(),
         { cors: true },
     );
+    const logger: Logger = new Logger("NestApplication", { timestamp: true });
     const configService: ConfigService = app.get(ConfigService);
     const PORT: string =  <string>configService.get<string>("PORT");
 
@@ -37,6 +38,6 @@ async function bootstrap()
     });
     app.useGlobalPipes(new ValidationPipe());
     await app.listen(PORT, "0.0.0.0");
-    console.log(`Application is running on: ${await app.getUrl()}`);
+    logger.log(`Application is running on: ${await app.getUrl()}`);
 }
 bootstrap();
