+++ swagger-spec.json
@@ -1,1 +1,1 @@
-{"openapi":"3.0.0","paths":{"/v1/pairs":{"get":{"operationId":"PairController_getPairs","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["pairs"]}},"/v1/pairs/{address}":{"get":{"operationId":"PairController_getPairFromAddress","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["pairs"]}},"/v1/tokens":{"get":{"operationId":"TokenController_getTokens","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tokens"]}},"/v1/tokens/{address}":{"get":{"operationId":"TokenController_getTokenFromAddress","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tokens"]}},"/v1/trading-competition":{"get":{"operationId":"TradingCompetitionController_getTradingCompetitionRanking","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["trading-competition"]}},"/health":{"get":{"operationId":"HealthController_getHealth","parameters":[],"responses":{"200":{"description":""}},"tags":["health"]}}},"info":{"title":"Vexchange API","description":"","version":"1.0.2","contact":{}},"tags":[],"servers":[],"components":{"schemas":{}}}
+{"openapi":"3.0.0","paths":{"/v1/pairs":{"get":{"operationId":"PairController_getPairs","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["pairs"]}},"/v1/pairs/{address}":{"get":{"operationId":"PairController_getPairFromAddress","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["pairs"]}},"/v1/tokens":{"get":{"operationId":"TokenController_getTokens","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tokens"]}},"/v1/tokens/{address}":{"get":{"operationId":"TokenController_getTokenFromAddress","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tokens"]}},"/v1/trading-competition/ranking":{"get":{"operationId":"TradingCompetitionController_getRanking","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["trading-competition"]}},"/health":{"get":{"operationId":"HealthController_getHealth","parameters":[],"responses":{"200":{"description":""}},"tags":["health"]}}},"info":{"title":"Vexchange API","description":"","version":"1.0.3","contact":{}},"tags":[],"servers":[],"components":{"schemas":{}}}
